Splitting SportClass into multiple classes (not yet working)
